Extract the Http boundary module and lock the flat Api kernel empty [#807]
The last three types in the flat SSO-Auth/Api/ root — SSOController, its
RequestHelpers, and the admin ProviderConnectionTester probe — move into a new
Api/Http module. With them gone the flat Api holds no source at all, so the
kernel dissolution the module migration set out to do (#777) is complete.

- Api/Http: the web boundary (namespace ...Api.Http). It is the composition top
  of the DAG — it imports Audit, Avatar, Flows, Linking, Net, Oidc, Provider,
  Saml, Session and Shared (wide by design, it fronts every flow) and nothing
  imports it back. Its DAG InlineData case declares exactly those edges.
- Every cross-module reference to the moved types was a doc cref only (proof the
  kernel held no code the modules depended on): OidcDiscoveryOptions,
  OidcLoginService, SamlLoginService, ProviderNameValidator and SsoRateLimitGate
  had `<see cref="SSOController"/>` / `ProviderConnectionTester` crefs, now plain
  text per the cross-module doc rule. No new code edge exists.
- New fitness function FlatApi_HoldsNoSourceFiles_EveryApiTypeLivesInAModule
  fails if any .cs reappears directly in Api/, locking the kernel dissolved.
- The two module-scoped helper tests (RequestHelpersTests,
  ProviderConnectionTesterTests) move to Tests/Http to keep the mirror; the
  controller's own split tests stay in Tests/Kernel (not module-scoped by the
  mirror rule).
- docs/ARCHITECTURE.md: the module map gains the Http/Identity/Routing/Session
  rows and corrected Saml/Oidc edges; the "shared kernel" section is replaced by
  "The kernel is dissolved".

Behaviour-preserving: the move is namespace + doc only (the persisted
AuthenticationProviderId was already decoupled from the controller type in
#837), verified by the full build (--warnaserror both TFMs), the suite (1592
both TFMs, +2 conformance rules) and the ABI-floor build.

Closes #807
Refs #790, #777, #791
