+++ SSO-Auth/Api/Http/RequestHelpers.cs
