+++ SSO-Auth/Api/Http/SSOController.cs
@@ -33,7 +33,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 
-namespace Jellyfin.Plugin.SSO_Auth.Api;
+namespace Jellyfin.Plugin.SSO_Auth.Api.Http;
 
 /// <summary>
 /// The sso api controller.
